Relocate the migrations fold to helpers and widen the AST guard walk
