+++ esphome_device_builder/helpers/artifacts_tarball_members.py
@@ -6,7 +6,11 @@ from __future__ import annotations
 # receiver-side packer (``controllers.remote_build.artifacts_tarball``)
 # writes them; the offloader-side materialiser
 # (``helpers.remote_artifacts_materialise``) pulls them out and stages
-# them at the offloader's canonical cache locations.
+# them at the offloader's canonical cache locations; the WS-adapter
+# (``unpack_artifacts_response``) ignores ``storage.json`` /
+# ``platformio.ini`` (they're not flash images) and reads
+# ``idedata.json`` to recover the upstream-canonical flash-image
+# manifest.
 STORAGE_MEMBER_NAME = "storage.json"
 IDEDATA_MEMBER_NAME = "idedata.json"
 PLATFORMIO_INI_MEMBER_NAME = "platformio.ini"
